% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnectorAuth.R
\docType{package}
\name{DatabaseConnectorAuth}
\alias{DatabaseConnectorAuth}
\title{DatabaseConnectorAuth}
\description{
This package provides Windows integrated authenication DLL for \code{sqljdbc42}
}
