% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnectorAuth.R
\name{getAuthDllPath}
\alias{getAuthDllPath}
\title{Get authenication DLL path for \code{sqljdbc42}}
\usage{
getAuthDllPath()
}
\description{
This function returns the package path to the Windows integrated authenication DLL
for \code{sqljbdc42}.
}
